requirements.txt
langchain
langchain-core
langchain-astradb
langchain-google-genai
langchain-graph-retriever[astra]
PyMuPDF
Pillow
numpy
pandas
python-dotenv
scikit-learn
fastapi
uvicorn
pinecone
google-generativeai
langchain-pinecone